+++ gear/form_qc_checker/src/python/form_qc_app/main.py
@@ -5,10 +5,13 @@ Uses nacc-form-validator (https://github.com/naccdata/nacc-form-
 validator) for validating the inputs.
 """
 
+import json
 import logging
+from json.decoder import JSONDecodeError
 from typing import Any, Dict, List, Optional
 
 from centers.nacc_group import NACCGroup
+from configs.ingest_configs import FormProjectConfigs, ModuleConfigs
 from flywheel import FileEntry
 from flywheel.rest import ApiException
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
@@ -18,7 +21,7 @@ from gear_execution.gear_execution import (
     GearExecutionError,
     InputFileWrapper,
 )
-from keys.keys import DefaultValues, FieldNames
+from keys.keys import DefaultValues
 from nacc_form_validator.quality_check import (
     QualityCheck,
     QualityCheckException,
@@ -101,6 +104,19 @@ def validate_input_file_type(mimetype: str) -> Optional[str]:
     return None
 
 
+def load_supplement_input(
+        supplement_input: InputFileWrapper) -> Optional[Dict[str, Any]]:
+    with open(supplement_input.filepath, mode='r',
+              encoding='utf-8') as file_obj:
+        try:
+            input_data = json.load(file_obj)
+        except (JSONDecodeError, TypeError) as error:
+            log.error('Failed to load supplement input file %s - %s',
+                      supplement_input.filename, error)
+            return None
+    return input_data
+
+
 def run(  # noqa: C901
         *,
         client_wrapper: ClientWrapper,
@@ -108,7 +124,9 @@ def run(  # noqa: C901
         s3_client: S3BucketReader,
         admin_group: NACCGroup,
         gear_context: GearToolkitContext,
-        redcap_connection: Optional[REDCapReportConnection] = None):
+        form_project_configs: FormProjectConfigs,
+        redcap_connection: Optional[REDCapReportConnection] = None,
+        supplement_input: Optional[InputFileWrapper] = None):
     """Starts QC process for input file. Depending on the input file type calls
     the appropriate file processor.
 
@@ -118,7 +136,9 @@ def run(  # noqa: C901
         s3_client: boto3 client for QC rules S3 bucket
         admin_group: Flywheel admin group
         gear_context: Flywheel gear context
-        redcap_connection (Optional): REDCap project for NACC QC checks
+        form_project_configs: module configurations
+        redcap_connection (optional): REDCap project for NACC QC checks
+        supplement_input (optional): input file for supplement module
 
     Raises:
         GearExecutionError if any problem occurs while validating input file
@@ -152,33 +172,47 @@ def run(  # noqa: C901
         raise GearExecutionError(
             f'Failed to find the project with ID {file.parents.project}')
 
-    legacy_label = gear_context.config.get('legacy_project_label',
-                                           DefaultValues.LEGACY_PRJ_LABEL)
-    pk_field = (gear_context.config.get('primary_key',
-                                        FieldNames.NACCID)).lower()
-    date_field = (gear_context.config.get('date_field',
-                                          FieldNames.DATE_COLUMN)).lower()
+    if (module not in form_project_configs.accepted_modules
+            or not form_project_configs.module_configs.get(module)):
+        raise GearExecutionError(
+            f'Failed to find the configurations for module {module}')
+
+    legacy_label = (form_project_configs.legacy_project_label
+                    if form_project_configs.legacy_project_label else
+                    DefaultValues.LEGACY_PRJ_LABEL)
+    pk_field = form_project_configs.primary_key.lower()
+    module_configs: ModuleConfigs = form_project_configs.module_configs.get(
+        module)  # type: ignore
+    date_field = module_configs.date_field
     strict = gear_context.config.get("strict_mode", True)
 
     error_writer = ListErrorWriter(container_id=file_id,
                                    fw_path=proxy.get_lookup_path(file))
 
     rule_def_loader = DefinitionsLoader(s3_client=s3_client,
-                                        strict=strict,
-                                        error_writer=error_writer)
+                                        error_writer=error_writer,
+                                        strict=strict)
 
     error_store = REDCapErrorStore(redcap_con=redcap_connection)
     gear_name = gear_context.manifest.get('name', 'form-qc-checker')
 
     file_processor: FileProcessor
     if file_type == 'json':
+        supplement_record = load_supplement_input(
+            supplement_input=supplement_input) if supplement_input else None
+        if module_configs.supplement_module and not supplement_record:
+            raise GearExecutionError(
+                f"Supplement {module_configs.supplement_module['label']} "
+                f"visit record is required to validate {module} visit")
+
         file_processor = JSONFileProcessor(pk_field=pk_field,
                                            module=module,
                                            date_field=date_field,
                                            project=ProjectAdaptor(
                                                project=project, proxy=proxy),
                                            error_writer=error_writer,
-                                           gear_name=gear_name)
+                                           gear_name=gear_name,
+                                           supplement_data=supplement_record)
     else:  # For enrollment form processing
         file_processor = CSVFileProcessor(pk_field=pk_field,
                                           module=module,