+++ docs/identifier_lookup/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.0.8
+* Fixes CSV line number in error reports (exclude header row)
+  
 ## 1.0.7
 * Adds pre-processing check for PTID length